requirements.txt
google-genai>=1.0.0
resend>=2.0.0
python-dotenv>=1.0.0
jinja2>=3.1.0
aiohttp>=3.9.0
httpx>=0.27.0
feedparser>=6.0.0
